# Perf ignore uuid and timestamp generation for NoopContext