--- a/pyproject.toml
+++ b/pyproject.toml
@@ -100,6 +100,7 @@ gunicorn = "*"
 uvicorn = "*"
 kafka-python = "*"
 requests = "*"
+pytest-benchmark = "4.0.0"
 
 [tool.poetry.group.plugins.dependencies]
 urllib3 = "1.26.7"